store/requirements.txt
asgiref==3.6.0
attrs==23.1.0
autopep8==2.0.2
Django==4.2.1
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
drf-spectacular==0.26.2
environ==1.0
flake8==6.0.0
gprof2dot==2022.7.29
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
Pillow==9.5.0
psycopg==3.1.9
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
